Update troubleshoot-pi-adapter-for-dnp3.md
I only listed parameters that I thought would be of interest during troubleshooting. For instance, the "Id" parameter for each of the configurations is a friendly name and is simple/redundant across all configs. Number of retries is self explanatory and should not be the cause of any issues, but I could add that to the list for completeness under the MasterStationBehavior section